requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
python-dotenv==1.0.1
groq==0.11.0
httpx==0.27.2
scrapling==0.4.3
pytest==8.3.4
pytest-asyncio==0.24.0
